pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "kautos-creation"
version = "0.1.0"
description = "A Python project"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Bill Dusch", email = "bill.dusch@gmail.com"},
]
dependencies = [
    "pandas",
    "notion-client",
    "python-dotenv",
    "anthropic",
    "pydantic",
    "streamlit"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 120
